Proper tabbing across items from listing (#41)
* Proper tabbing across items from listing

With this commit, the listing sets a tabindex value of 0 by default to
all elements in the list, but checkboxes for row selection. Also, sets
a value of tabindex=-1 to those items listing's with "tabindex" set to
"disabled".

Before this commit, the system was automatically generating sequential
tabindexes to all items from the listing and adding a gap of 100 between
rows. With this approach, proper tabbing when the view contains multiple
listings was not possible.

* Changelog

* Remove obsolete tabindex_start

* Remove unnecessary curly brackets